TOcean::WriteTo: keep the int loop counter
The short-counter change belongs to ReadFrom only -- applying it to WriteTo as
well cost 77.4% -> 69.6%, so the writer keeps its int counter and the comment
explaining the reader's 16-bit bound test stays with the reader. Writer back at
77.4%, reader unchanged at 47.2%.

+++ src/game/navy/TOcean.cpp
@@ -217,9 +217,7 @@ void TOcean::ReadFrom(TStream* stream) {
 void TOcean::WriteTo(TStream* stream) {
   TObject::WriteTo(stream);
   stream->WriteBytes(&nationCount, 2);
-  // A short counter: the bound test is a 16-bit `cmp word ptr [eax], bx` (0x5623a3), so
-  // an int loop variable would need a movsx of nationCount on every compare.
-  short i;
+  int i;
   for (i = 0; i < nationCount; ++i) {
     contextArray[i].WriteTo(stream);
   }
